requirements.txt
flask==3.1.0
groq==0.11.0
python-dotenv==1.0.1
gunicorn==23.0.0
